Completed UI for guest

+++ src/main/java/darien/venusrestblog/controllers/PostsController.java
@@ -3,13 +3,18 @@ package darien.venusrestblog.controllers;
 import darien.venusrestblog.data.Category;
 import darien.venusrestblog.data.Post;
 import darien.venusrestblog.data.User;
+import darien.venusrestblog.data.UserRole;
+import darien.venusrestblog.misc.FieldHelper;
 import darien.venusrestblog.repository.CategoriesRepository;
 import darien.venusrestblog.repository.PostsRepository;
 import darien.venusrestblog.repository.UsersRepository;
 import lombok.AllArgsConstructor;
+import org.springframework.beans.BeanUtils;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.web.bind.annotation.*;
-
+import org.springframework.web.server.ResponseStatusException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -32,13 +37,21 @@ public class PostsController {
         return postsRepository.findById(id);
     }
 
-    @PostMapping(path = "")
-    private void createPost(@RequestBody Post newPost, OAuth2Authentication auth){
-        newPost.setCategories(new ArrayList<>());
+    @PostMapping("")
+    @PreAuthorize("hasAuthority('USER') || hasAuthority('ADMIN')")
+    public void createPost(@RequestBody Post newPost, OAuth2Authentication auth) {
+        if(newPost.getTitle() == null || newPost.getTitle().length() < 1) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Title cannot be blank!");
+        }
+        if(newPost.getContent() == null || newPost.getContent().length() < 1) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Content cannot be blank!");
+        }
 
-        String username = auth.getName();
-        User author = usersRepository.findByUsername(username);
+        String userName = auth.getName();
+        User author = usersRepository.findByUsername(userName);
         newPost.setAuthor(author);
+
+        newPost.setCategories(new ArrayList<>());
 
         // use first 2 categories for the post by default
         Category cat1 = categoriesRepository.findById(1L).get();
@@ -51,14 +64,40 @@ public class PostsController {
     }
 
     @PutMapping("/{id}")
+    @PreAuthorize("hasAuthority('USER') || hasAuthority('ADMIN')")
     public void updatePost(@RequestBody Post updatedPost, @PathVariable long id) {
-        // in case id  is not in req body set it with path var id
+        Optional<Post> originalPost = postsRepository.findById(id);
+        if(originalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post " + id + " not found");
+        }
+
+        // in case id is not in the request body (i.e., updatedPost), set it
+        // with the path variable id
         updatedPost.setId(id);
-        postsRepository.save(updatedPost);
+
+        // copy any new field values FROM updatedPost TO originalPost
+        BeanUtils.copyProperties(updatedPost, originalPost.get(), FieldHelper.getNullPropertyNames(updatedPost));
+
+        postsRepository.save(originalPost.get());
     }
 
-    @DeleteMapping ("/{id}")
-    public void DeletePostById(@PathVariable long id) {
+    @DeleteMapping("/{id}")
+    @PreAuthorize("hasAuthority('USER') || hasAuthority('ADMIN')")
+    public void deletePostById(@PathVariable long id, OAuth2Authentication auth) {
+        String userName = auth.getName();
+        User loggedInUser = usersRepository.findByUsername(userName);
+
+        Optional<Post> optionalPost = postsRepository.findById(id);
+        if(optionalPost.isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Post id " + id + " not found");
+        }
+        // grab the original post from the optional and check the logged in user
+        Post originalPost = optionalPost.get();
+
+        // admin can delete anyone's post. author of the post can delete only their posts
+        if(loggedInUser.getRole() != UserRole.ADMIN && originalPost.getAuthor().getId() != loggedInUser.getId()) {
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "Not yo post!");
+        }
         postsRepository.deleteById(id);
     }
 }
